+++ tests/TestRPC.hs
@@ -26,6 +26,8 @@ import Network.Transport.Memory
 
 import Control.Concurrent
 
+import qualified Data.Map as M
+
 import Test.Framework
 import Test.HUnit
 import Test.Framework.Providers.HUnit
@@ -41,7 +43,8 @@ tests = [
     testCase "call-one-handler" testOneHandler,
     testCase "call-two-handlers" testTwoHandlers,
     testCase "gcall-three-handlers" testGroupCall,
-    testCase "call-one-with-timeout" testOneHandlerWithTimeout
+    testCase "call-one-with-timeout" testOneHandlerWithTimeout,
+    testCase "gcall-three-handlers-with-timeout"testGroupCallWithTimeout
   ]
 
 testOneHandler :: Assertion
@@ -100,11 +103,11 @@ testGroupCall = do
     h3 <- handle endpoint3 name3 "foo" $ \msg -> if msg == "hello" then return "bar" else return ""
     h4 <- handle endpoint4 name4 "foo" $ \msg -> if msg == "hello" then return "baz" else return ""
     let cs = newCallSite endpoint1 name1
-    results <- gcall cs [name2,name3,name4] "foo" "hello"
-    assertBool "Foo not present in results" (elem "foo" results)
-    assertBool "Bar not present in results" (elem "bar" results)
-    assertBool "Bar not present in results" (elem "baz" results)
-    assertEqual "Unxpected number of results" 3 (length results)
+    results <- (gcall cs [name2,name3,name4] "foo" "hello") :: IO (M.Map Name String)
+    assertBool "Foo not present in results" (elem "foo" $ M.elems results)
+    assertBool "Bar not present in results" (elem "bar" $ M.elems results)
+    assertBool "Bar not present in results" (elem "baz" $ M.elems results)
+    assertEqual "Unxpected number of results" 3 (M.size results)
     hangup h2
     hangup h3
     hangup h4
@@ -136,4 +139,40 @@ testOneHandlerWithTimeout = do
     result2 <- (callWithTimeout cs2 name2 "foo" shorter "hello") :: IO (Maybe String)
     assertEqual "Result not expected value" Nothing result2
     hangup h2
-    
+
+testGroupCallWithTimeout :: Assertion
+testGroupCallWithTimeout = do
+    let name1 = "endpoint1"
+        name2 = "endpoint2"
+        name3 = "endpoint3"
+        name4 = "endpoint4"
+        longest = 750 * 1000 -- three quarters of a second
+        longer = 500 * 1000 -- half a second
+        shorter = 250 * 1000 -- quarter second
+    transport <- newMemoryTransport
+    endpoint1 <- newEndpoint [transport]
+    endpoint2 <- newEndpoint [transport]
+    endpoint3 <- newEndpoint [transport]
+    endpoint4 <- newEndpoint [transport]
+    Right () <- bindEndpoint endpoint1 name1
+    Right () <- bindEndpoint endpoint2 name2
+    Right () <- bindEndpoint endpoint3 name3
+    Right () <- bindEndpoint endpoint4 name4
+    h2 <- handle endpoint2 name2 "foo" $ \msg -> do
+        threadDelay shorter
+        if msg == "hello" then return "foo" else return ""
+    h3 <- handle endpoint3 name3 "foo" $ \msg ->  do
+        threadDelay shorter
+        if msg == "hello" then return "bar" else return ""
+    h4 <- handle endpoint4 name4 "foo" $ \msg ->  do
+        threadDelay longest
+        if msg == "hello" then return "baz" else return ""
+    let cs = newCallSite endpoint1 name1
+    results <- gcallWithTimeout cs [name2,name3,name4] "foo" longer "hello"
+    assertEqual "Foo not present in results" (Just $ Just "foo") (M.lookup name2 results)
+    assertEqual "Bar not present in results" (Just $ Just "bar") (M.lookup name3 results)
+    assertEqual "Baz shouldn't be present in results" (Just $ Nothing) (M.lookup name4 results)
+    assertEqual "Unxpected number of results" 3 (M.size results)
+    hangup h2
+    hangup h3
+    hangup h4